add created news details page

+++ app/controllers/CreatedNewsController.js
@@ -1,16 +1,18 @@
 'use strict';
 
-newsApp.controller('CreatedNewsController', function($scope, localStorageService) {
+newsApp.controller('CreatedNewsController', function($scope, $routeParams, localStorageService) {
 
-    let getCreatedNews = () => {
-        $scope.createdNews = localStorageService.get('createdNews');
-    };
-
-    getCreatedNews();
-
-    $scope.deleteCreatedNews = (title) => {
-        localStorageService.deleteCreatedNews(title);
+    if ($routeParams['title']) {
+        $scope.newsDetails = localStorageService.getCreatedNewsDetails($routeParams['title']);
+    } else {
+        let getCreatedNews = () => {
+            $scope.createdNews = localStorageService.get('createdNews');
+        };
         getCreatedNews();
-    };
+        $scope.deleteCreatedNews = (title) => {
+            localStorageService.deleteCreatedNews(title);
+            getCreatedNews();
+        };
+    }
 
 });
